requirements.txt
asgiref==3.7.2
astroid==3.0.1
autopep8==2.0.4
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
dill==0.3.7
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==4.2.7
django-phonenumber-field==7.2.0
django-select-multiple-field==0.4.2
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.2
pep8==1.7.1
phonenumberslite==8.13.25
platformdirs==4.0.0
psycopg2==2.9.9
pycodestyle==2.11.1
pylint==3.0.2
pylint-flask==0.6
pylint-plugin-utils==0.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
